pyproject.toml
[tool.poetry]
name = "awd-cmd"
version = "0.1.0"
description = ""
authors = ["Yoshino-s <yoshino.prog@gmail.com>"]
readme = "README.md"
packages = [{include = "awd_cmd"}]

[tool.poetry.dependencies]
python = "^3.11"
tabulate = "^0.9.0"
click = "^8.1.3"
paramiko = "^3.2.0"
pyyaml = "^6.0"
verboselogs = "^1.7"
httpx = "^0.24.1"
pydantic = "^1.10.9"
coloredlogs = "^15.0.1"
rich = "^13.4.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
pre-commit = "^3.3.2"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
